# btf: optimize Spec.Copy