btf: avoid allocating typeDeque in copier.copy
    
        goos: linux
        goarch: amd64
        pkg: github.com/cilium/ebpf/btf
        cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                    │   old.txt   │              new.txt               │
                    │   sec/op    │   sec/op     vs base               │
        SpecCopy-16   37.76m ± 2%   33.27m ± 5%  -11.91% (p=0.002 n=6)
    
                    │   old.txt    │               new.txt               │
                    │     B/op     │     B/op      vs base               │
        SpecCopy-16   38.10Mi ± 0%   32.28Mi ± 0%  -15.27% (p=0.002 n=6)
    
                    │   old.txt   │              new.txt               │
                    │  allocs/op  │  allocs/op   vs base               │
        SpecCopy-16   294.7k ± 0%   203.2k ± 0%  -31.05% (p=0.002 n=6)
    
    Signed-off-by: Lorenz Bauer <lmb@isovalent.com>

btf: pre-size copier
    
    Make sure copier has the appropriate size from the start.
    
        goos: linux
        goarch: amd64
        pkg: github.com/cilium/ebpf/btf
        cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                    │   old.txt   │              new.txt               │
                    │   sec/op    │   sec/op     vs base               │
        SpecCopy-16   43.96m ± 8%   37.76m ± 2%  -14.10% (p=0.002 n=6)
    
                    │   old.txt    │               new.txt               │
                    │     B/op     │     B/op      vs base               │
        SpecCopy-16   48.44Mi ± 0%   38.10Mi ± 0%  -21.36% (p=0.002 n=6)
    
                    │   old.txt   │              new.txt              │
                    │  allocs/op  │  allocs/op   vs base              │
        SpecCopy-16   299.3k ± 0%   294.7k ± 0%  -1.56% (p=0.002 n=6)
    
    Signed-off-by: Lorenz Bauer <lmb@isovalent.com>

btf: add Spec.Copy benchmark
    
    Signed-off-by: Lorenz Bauer <lmb@isovalent.com>